Iterable for file of concatenated XML  documents.
This is handy for when you must process all of Gigaword, because its much
faster to process all the documents of a month concatenated together
than to open them all individually.

This should get refactored to not be an Iterable, though, because as it
is we lose control over closing the file.