Support KB ids in EDLMention
Changes the EDL classes to support an optional KB identifier, with a
helper for creating NIL KB IDs. Code depending on EDLWriter.create()
will continue to see the old behavior of writing `NIL` for the KB ID
of all EDL mentions, but newer code that uses EDLWriter.Builder will
be able to specify a default KB ID to be used for any EDL mentions
with absent KB IDs.